Adding parser module, README, and a test_driver folder.
Parser includes contributions from Jared, Keeler, and Corbin,
many of which have not been pushed to the repo until this point.
Jared's earlier code, related to the sense_body parser, has been
merged with this new parser module. Changes were made for the sake
of consistency, as well as for compatibility with Visual Studio.

README details which libraries to include to compile (need wsock32.lib).

The test_drivers folder should contain any further testing code,
the file in this commit is parse_test.cpp, which runs a few tests
on the parser and prints the results. Fairly simple, but if you have
this lying around in the same directory as a program with another main()
definition, you'll get a linker error.