Integration of threaded UDP client code with parsing code.
Added a Player class, which accepts a string from the client
and passes it to the parser to parse. The Player class will, in
the future, will implement AI functionality as well.

Also silenced warnings having to do with double to float conversions
by changing the type definitions in the Parser.hpp structs to
doubles rather than floats.

Changed the printServerHash function in Player.cpp slightly to avoid
any problems with it. Joe mentioned that it was causing his build to
crash, but I haven't had any problems with it.

-- Keeler