Added thread-safe and memory-safe parsing functions.
The strtok() function was used widely throughout our parser,
but it (along with some other string functions), were throwing
warnings that they were unsafe. In addition, I found that strtok()
is not thread safe because it stores its parsing context in a
static variable which is clobbered each time it is called. To rid
us of compiler warnings, as well as make the code thread safe, I
used strtok_s() instead, which stores the parsing context as a local
variable. Other functions like strcpy() were replaced with safe
versions as well.

Additionally, I added a constructor to the Player class, so that it
will initialize the aural and sense_body structs to invalid values,
rather than the garbage values that they were before. This makes the
file printing look a lot less ugly.