Adding: stationary flags, memory queues, and a Vector class.
These changes were added before in separate pushes to the repo, but were
overwritten accidentally. So here they are in their entirety:

1) Vector class: will use this for representing 2d coordinates, as well as
2d velocities. Specifically for absolute position estimates, etc.

2) Stationary flags: a hash table containing the absolute position of each
of the stationary flags on the field. The values have not been tested yet,
and so they are approximate. Will refine them later, but they'll do for now,
since Corbin needs them to work on the position estimation code.

3) Memory: put the sensory data parsed from the server in a series of queues
corresponding to the type of data parsed. This is a form of "memory" for the
client, so that it can recall past messages and object locations for a
finite period of time. Once the queue is full, it will push the oldest item
off when you add a new item on. Each item is like a snapshot of that
timestamp.