Added location estimation for nonvisible objects.
Major change: using deque instead of queue for player memory. This is due to
the fact that the deque class has all of the functionality of the queue
class, except that it also includes iterators and constant array-like access
using the [] operators, which the queue class does not.

Note that I haven't implemented this functionality for other players yet,
I'm still figuring out how I want to do that. I also fixed a small error in
udp_client.hpp so that the STL queue class would be included.

Also, the estimation is done linearly, with no regards to friction. The
server models friction on the ball, so it would be more accurate to include
that model.