Added functionality for determing absolute position and velocity.
Added a function to determine if a player is open for passing, a function
that returns the absolute value of two angles relative to the x-axis, a
function that builds a vector of visible, in range teammates,
opponents, or unidentified players, and a function that converts each
visible object's relative position and velocity into absolute positions and
velocities with respect to the center of the field. Also made slight changes
to main.cpp, made a private teamName member in the Player class and modified
the Player::parseBuffer function to accomodate the conversions.