Added player role setter function.
Sets the player's role, or position, that determines how the player should
behave on the field. For example, goalie, defender, midfielder, forward.

I tried doing this with an enum statement placed in ai_processing.hpp, but
ran into some dumb errors and decided to go with a string to represent the
role.