Changed think function to work with Joe's transmission queue.
See udp_client.cpp for how it's used. Only one turn/dash/kick can be sent
every 100ms still, but now you can queue up "say" of "turn neck" messages to
be sent alongside those turn/dash/kick commands. If you queue up a turn and
a kick in one queue, both will be sent but the server will ignore the kick
because it was sent within the same 100ms as the turn.