Altered when commands are sent, and fixed goalie positioning logic.
Changed code so that think happens in transmit function right before
sending, rather than un the completion routine. This allows the client to
send based on the most recent data.

Goalie still needs plenty of work, but I've got it turning his head to keep
his eye on the ball while dashing to a point. There are a few problems with
him turning a little too far and not seeing the ball anymore, causing him to
enter different logic functions. I'm going to try and make it so he can keep
his head turned exactly (or as close as that gets with this server) at the
ball at all times while dashing.