Made goalie constantly try to position himself between ball and goal.
I made a function to calculate the point that the goalie should go to given
a certain radius from the goal. The goalie should move to the point of
intersection between the line from ball to goal and the circle with its
center at the goalie's goal of the specified radius.

Also fixed some weird visual parser error; got some unexpected input when I
pitted Jared's forward against my goalie, couldn't find it in the docs so I
just ignore it.